requirements.txt
langchain
streamlit
langchain_ollama
python-dotenv